requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.4
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
filelock==3.13.4
identify==2.5.36
idna==3.7
inflection==0.5.1
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
platformdirs==4.2.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==69.5.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.3
